Rename workflowrecord to apprecord
