+++ client/public/js/records/app_record.js
@@ -1,14 +1,14 @@
 import { Record, List as IList } from 'immutable';
 import RunRecord from './run_record';
 
-const WorkflowRecord = new Record({
+const AppRecord = new Record({
   bgColor: '',
   bgIndex: null,
   color: null,
   comingSoon: false,
   creatorImage: null,
   id: null,
-  fetching: false, // for fetching the workflow itself
+  fetching: false, // for fetching the app itself
   fetchingError: null,
   selectLigands: false,
   run: new RunRecord(),
@@ -16,8 +16,7 @@ const WorkflowRecord = new Record({
   tasks: new IList(),
   title: '',
   viewCount: 0,
-  workflowNodes: new IList(),
   description: 'no description provided',
 });
 
-export default WorkflowRecord;
+export default AppRecord;
